have authors section use its own tag catalog too
