--- packages/react-tei/src/tags/Name.tsx
@@ -1,11 +1,0 @@
-import Typography from "@mui/material/Typography";
-import type { ComponentProps } from "./type.js";
-import { Value } from "./Value.js";
-
-export function Name({ data }: ComponentProps) {
-	return (
-		<Typography>
-			<Value data={data.value} />
-		</Typography>
-	);
-}
